Compiler warning fix

+++ src/lib/ioloop-notify-inotify.c
@@ -61,7 +61,7 @@ static bool inotify_input_more(struct ioloop *ioloop)
 			break;
 
 		event = (struct inotify_event *)(event_buf + pos);
-		i_assert(event->len < ret);
+		i_assert(event->len < (size_t)ret);
 		pos += sizeof(*event) + event->len;
 
 		io = io_notify_fd_find(&ctx->fd_ctx, event->wd);
